+++ internal/ceres/block_sparse_matrix.h
@@ -79,7 +79,7 @@ class BlockSparseMatrix : public SparseMatrix {
 
   int num_rows()         const final { return num_rows_;     }
   int num_cols()         const final { return num_cols_;     }
-  int num_nonzeros()     const final { return num_nonzeros_; }
+  int64_t num_nonzeros()     const final { return num_nonzeros_; }
   const double* values() const final { return values_.get(); }
   double* mutable_values()     final { return values_.get(); }
