+++ internal/ceres/linear_operator.h
@@ -51,7 +51,7 @@ class LinearOperator {
 
   virtual int num_rows() const = 0;
   virtual int num_cols() const = 0;
-  virtual int num_nonzeros() const = 0;
+  virtual int64_t num_nonzeros() const = 0;
 };
 
 }  // namespace internal
