+++ internal/ceres/schur_jacobi_preconditioner.cc
@@ -100,7 +100,7 @@ void SchurJacobiPreconditioner::RightMultiply(const double* x,
 
 int SchurJacobiPreconditioner::num_rows() const { return m_->num_rows(); }
 
-int SchurJacobiPreconditioner::num_nonzeros() const {
+int64_t SchurJacobiPreconditioner::num_nonzeros() const {
   return m_->num_nonzeros();
 }
