+++ internal/ceres/sparse_cholesky.h
@@ -107,7 +107,7 @@ class SparseCholesky {
       double* solution,
       std::string* message);
 
-  virtual int num_nonzeros() const = 0;
+  virtual int64_t num_nonzeros() const = 0;
 
 };
 
@@ -127,7 +127,7 @@ class RefinedSparseCholesky : public SparseCholesky {
   virtual LinearSolverTerminationType Solve(const double* rhs,
                                             double* solution,
                                             std::string* message);
-  int num_nonzeros() const final { return sparse_cholesky_->num_nonzeros() + lhs_->num_nonzeros(); }
+  int64_t num_nonzeros() const final { return sparse_cholesky_->num_nonzeros() + lhs_->num_nonzeros(); }
 
  private:
   std::unique_ptr<SparseCholesky> sparse_cholesky_;
