+++ internal/ceres/suitesparse.h
@@ -296,7 +296,7 @@ class SuiteSparseCholesky : public SparseCholesky {
   LinearSolverTerminationType Solve(const double* rhs,
                                     double* solution,
                                     std::string* message) final;
-  int num_nonzeros() const final {
+  int64_t num_nonzeros() const final {
     return num_nonzeros_;
   };
 
@@ -306,7 +306,7 @@ class SuiteSparseCholesky : public SparseCholesky {
   const OrderingType ordering_type_;
   SuiteSparse ss_;
   cholmod_factor* factor_;
-  int num_nonzeros_;
+  int64_t num_nonzeros_;
 };
 
 }  // namespace internal
