+++ internal/ceres/triplet_sparse_matrix.h
@@ -70,7 +70,7 @@ class TripletSparseMatrix : public SparseMatrix {
   void ToTextFile(FILE* file) const final;
   int num_rows()        const final   { return num_rows_;     }
   int num_cols()        const final   { return num_cols_;     }
-  int num_nonzeros()    const final   { return num_nonzeros_; }
+  int64_t num_nonzeros()    const final   { return num_nonzeros_; }
   const double* values()  const final { return values_.get(); }
   double* mutable_values() final      { return values_.get(); }
   void set_num_nonzeros(int num_nonzeros);
